qbt_add/requirements.txt
requests>=2.32.2
PyYAML>=6.0.2
